ai/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0
protobuf>=5.0.0
anthropic>=0.40.0
redis>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-json-logger>=3.2.0
